split compileFn and mapper on different modules
take into account query option
